requirements.txt
lime>=0.2.0
numpy>=1.21.0
pandas>=2.0.0
scikit-learn>=0.24.2
torch>=1.9.0
transformers>=4.11.0
sqlalchemy>=1.4.0
pymysql>=1.0.2
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.0
anthropic>=0.8.0
mcp>=0.1.0
httpx>=0.24.0
openai>=1.0.0
sqlalchemy>=2.0.0
sqlparse>=0.4.0
numpy>=1.24.0
langchain-anthropic>=0.1.0
langchain-deepseek>=0.0.1
langgraph>=0.0.10
langchain-mcp-adapters
smithery>=0.1.0
nest-asyncio>=1.5.0
pyyaml>=6.0.0 
openai-agents
